Support for dithering using interleaved gradient noise
