+++ src/scene/constants.js
@@ -1023,9 +1023,17 @@ export const DITHER_NONE = 'none';
 export const DITHER_BAYER8 = 'bayer8';
 
 /**
- * Opacity is dithered using a blue noise texture.
+ * Opacity is dithered using a blue noise.
  *
  * @type {string}
  * @category Graphics
  */
 export const DITHER_BLUENOISE = 'bluenoise';
+
+/**
+ * Opacity is dithered using an interleaved gradient noise.
+ *
+ * @type {string}
+ * @category Graphics
+ */
+export const DITHER_IGNNOISE = 'ignnoise';
